functions/requirements.txt
functions-framework==3.3.0
google-cloud-aiplatform[pipelines]==1.22.1
